Implementación de Get, Post y Delete de centros, clases, profesores y alumnos

+++ frontend/src/app/layouts/admin-layout/centros/centros.component.ts
@@ -1,5 +1,6 @@
 import { Component, AfterViewInit  } from '@angular/core';
 import { CentroService } from '../../../services/centros.service';
+import Swal from 'sweetalert2'
 
 
 @Component({
@@ -7,7 +8,7 @@ import { CentroService } from '../../../services/centros.service';
   templateUrl: './centros.component.html',
   styleUrl: './centros.component.css'
 })
-export class CentrosComponent implements AfterViewInit  {
+export class CentrosComponent implements AfterViewInit {
 
   centrosData: any;
 
@@ -30,8 +31,25 @@ export class CentrosComponent implements AfterViewInit  {
   }
 
   eliminarCentro(id: number){
-    this.centroService.deleteCentro(id).subscribe(res => {
-      this.getCentros();
-    })
+    Swal.fire({
+      title: "¿Estás seguro?",
+      text: "Esta acción no se podrá deshacer",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Eliminar"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.centroService.deleteCentro(id).subscribe(res => {
+          this.getCentros();
+        })
+        Swal.fire({
+          title: "Centro Eliminado",
+          icon: "success"
+        });
+      }
+    });
   }
+
 }
